requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
pymysql
python-dotenv
cryptography
